main/raspberry_pi/pyproject.toml
[tool.poetry]
name = "raspberry-pi"
version = "0.1.0"
description = ""
authors = ["kiinami <kinami@posteo.net>"]

[tool.poetry.dependencies]
python = "^3.11,<3.13"
adafruit-circuitpython-rgb-display = "^3.12.4"
pillow = "==9.5.0"
rpi-lgpio = "^0.6"
spidev = "^3.6"
numpy = "^1.26.4"
st7735 = "^1.0.0"
gpiod = "^2.1.3"
gpiodevice = "^0.0.3"
pyserial = "^3.5"
sysv-ipc = "^1.1.0"
audioio = "^2.1.0"
adafruit-circuitpython-hcsr04 = "^0.4.18"
python-statemachine = "^2.1.2"
pygame = "^2.5.2"
rpi-ws281x = "^5.0.0"
adafruit-circuitpython-neopixel = "^6.3.11"
pylance = "^0.11.0"
rich = "^13.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
